test: merge test_securityheaders.py into test_security_headers.py (#1013)
Both files covered config.Security; the reporting/CSP-directive/extend_csp
test classes from develop now live alongside the rest.
